@@ -1,17 +1,20 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 31
-    buildToolsVersion "31.0.0"
+    compileSdk 34
+    namespace 'com.crystal.crystalrangeseekbar'
 
     defaultConfig {
         applicationId "com.crystal.crystalrangeseekbar"
-        minSdkVersion 17
-        targetSdkVersion 31
+        minSdkVersion 21
+        targetSdk 34
         versionCode 1
         versionName "1.0"
     }
     buildTypes {
+        debug {
+
+        }
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
@@ -22,6 +25,18 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation project(':crystalrangeseekbar')
-    implementation 'com.android.support:appcompat-v7:27.0.2'
-    implementation 'com.android.support:design:27.0.2'
+
+    def fragment_version = "1.8.1"
+
+    // Java language implementation
+    implementation "androidx.fragment:fragment:$fragment_version"
+    // Kotlin
+    implementation "androidx.fragment:fragment-ktx:$fragment_version"
+    // Testing Fragments in Isolation
+    debugImplementation "androidx.fragment:fragment-testing-manifest:$fragment_version"
+    androidTestImplementation "androidx.fragment:fragment-testing:$fragment_version"
+
+    // https://mvnrepository.com/artifact/com.google.android.material/material
+    implementation group: 'com.google.android.material', name: 'material', version: '1.12.0'
+
 }
