@@ -1,17 +1,18 @@
 package com.crystal.crystalrangeseekbar.ui.activities;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.design.widget.TabLayout;
-import android.support.v4.app.Fragment;
-import android.support.v4.view.ViewPager;
-import android.support.v7.app.AppCompatActivity;
 
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.Fragment;
+import androidx.viewpager.widget.ViewPager;
 
 import com.crystal.crystalrangeseekbar.R;
 import com.crystal.crystalrangeseekbar.adapters.RangeSeekbarPagerAdapter;
 import com.crystal.crystalrangeseekbar.ui.fragments.RangeSeekbar;
 import com.crystal.crystalrangeseekbar.ui.fragments.Seekbar;
+import com.google.android.material.tabs.TabLayout;
 
 import java.util.ArrayList;
 import java.util.List;
